requirements.txt
pydim3
pyzmq
click
urwid
termcolor
numpy
